backend/AI_Launch_Pad/requirements.txt
Django==4.2.14
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-extensions==3.2.3
python-decouple==3.8
psycopg2-binary==2.9.7
redis==5.0.0
django-redis==5.3.0
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1

# AI Integration
google-generativeai==0.3.1
openai==1.3.5
tiktoken==0.5.1

# Social Media APIs
tweepy==4.14.0
python-twitter==3.5

# Data Processing & Analytics
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2

# Rate Limiting & Caching
django-ratelimit==4.1.0
django-cache-utils==0.7.2

# Monitoring & Logging
sentry-sdk==1.38.0
django-prometheus==2.3.1

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
responses==0.24.1

# Development
black==23.11.0
flake8==6.1.0
isort==5.12.0
pre-commit==3.5.0

# API Documentation
drf-spectacular==0.26.5
